refactor: 스피너 로직 수정
